refactor: Consolidate ledger to single binary format and restructure project
Remove all v1/v2 ledger code and documentation — the binary format is now
the only ledger specification. Rename ledger3* types/files to their canonical
names (Ledger, LedgerConfig, Header, Record, etc.) and set version byte to
0x01 since this is the first public release.

Move pkg/ to internal/orchestrator/ following standard Go project layout.
Update go.mod to Go 1.26.3.
