+++ .mise.toml
@@ -1,2 +1,3 @@
 [tools]
 node = "latest"
+go = "latest"
