+++ internal/orchestrator/ext_bazel.go
@@ -1,4 +1,4 @@
-package warden
+package orchestrator
 
 import (
 	"encoding/pem"
