+++ internal/orchestrator/ext_pip.go
@@ -1,4 +1,4 @@
-package warden
+package orchestrator
 
 type ExtPip struct{}
