+++ internal/orchestrator/ext_truststore.go
@@ -1,4 +1,4 @@
-package warden
+package orchestrator
 
 import (
 	"fmt"
