+++ internal/orchestrator/orchestrator.go
@@ -1,4 +1,4 @@
-package warden
+package orchestrator
 
 import (
 	"bufio"
