+++ relay/ledger.go
@@ -2,256 +2,265 @@ package relay
 
 import (
 	"crypto/ed25519"
-	"crypto/md5"
+	crypto_md5 "crypto/md5"
 	"crypto/rand"
-	"crypto/sha1"
-	"crypto/sha256"
+	crypto_sha1 "crypto/sha1"
+	crypto_sha256 "crypto/sha256"
 	"crypto/sha512"
-	"encoding/base64"
 	"encoding/binary"
-	"encoding/hex"
-	"encoding/json"
-	"encoding/pem"
 	"fmt"
 	"hash"
 	"io"
 	"log"
 	"sync/atomic"
-	"time"
 
+	"github.com/fxamacker/cbor/v2"
 	"golang.org/x/crypto/blake2b"
 )
 
-// Ledger implements the BuildWarden ledger v2 specification.
+// Record type byte values per the BuildWarden Ledger Specification.
+const (
+	RecordOpen       byte = 0x01
+	RecordCheckpoint byte = 0x02
+	RecordClose      byte = 0x03
+	RecordArtifact   byte = 0x04
+)
+
+// SchemaNoMetadata indicates no metadata is attached to a record.
+const SchemaNoMetadata byte = 0xFF
+
+// Ledger implements the BuildWarden binary ledger specification.
 // All writes are serialized through a single channel.
 type Ledger struct {
-	key         ed25519.PrivateKey
-	writer      io.Writer
-	enc         *json.Encoder
-	entries     chan entryRequest
-	done        chan struct{}
-	seq         atomic.Int64
-	hashes      []string
-	headerSig   string
-	prevSigRaw  []byte // cached decoded bytes of the previous signature
-	sigBuf      []byte // reusable buffer for signature input construction
+	key           ed25519.PrivateKey
+	writer        io.Writer
+	entries       chan entryRequest
+	done          chan struct{}
+	seq           atomic.Int64
+	hashes        []string
+	sigSize       int
+	hashBlockSize int
+	prevSigRaw    []byte
+	sigBuf        []byte // reusable buffer for signature input
 }
 
-// entryRequest is the union type sent through the single channel.
 type entryRequest struct {
 	entry  entryInput
-	// nil for fire-and-forget (checkpoint/close), non-nil for synchronous open
-	result chan<- string
+	result chan<- []byte // nil for fire-and-forget; returns raw signature for open
 }
 
 type entryInput struct {
-	Type          string
-	OpenSignature string // for checkpoint/close
-	Direction     string // for checkpoint/close
-	Size          int64  // payload size
-	PreHashed     map[string]string // pre-computed hashes
-	Metadata      map[string]any
+	Type        byte
+	OpenSig     []byte // for checkpoint/close/artifact
+	PayloadSize int64  // signed: negative=out, positive=in, 0=none
+	HashBlock   []byte // raw concatenated hashes (nil when payload=0)
+	SchemaIndex byte
+	Metadata    []byte // pre-encoded CBOR (nil when SchemaIndex=0xFF)
 }
 
-// LedgerEntry is the JSON structure written to the ledger file.
-type LedgerEntry struct {
-	EntryType     string         `json:"entry_type"`
-	OpenSignature string         `json:"open_signature,omitempty"`
-	Direction     string         `json:"direction,omitempty"`
-	Payload       *PayloadRecord `json:"payload,omitempty"`
-	Signature     string         `json:"signature"`
-	Seq           int64          `json:"seq"`
-	Timestamp     string         `json:"timestamp"`
-	Metadata      map[string]any `json:"metadata,omitempty"`
+// HeaderMeta is the CBOR metadata written after the header signature.
+type HeaderMeta struct {
+	Hashes      []string       `cbor:"hashes"`
+	Schemas     []string       `cbor:"schemas"`
+	Environment map[string]any `cbor:"environment"`
 }
 
-// HeaderEntry is the JSON structure for the ledger header.
-type HeaderEntry struct {
-	EntryType       string         `json:"entry_type"`
-	Version         string         `json:"version"`
-	Format          string         `json:"format"`
-	SignatureScheme string         `json:"signature_scheme"`
-	Hashes          []string       `json:"hashes"`
-	Environment     map[string]any `json:"environment"`
-	Payload         *PayloadRecord `json:"payload"`
-	Signature       string         `json:"signature"`
+// LedgerConfig holds parameters for creating a new ledger.
+type LedgerConfig struct {
+	Writer      io.Writer
+	Environment map[string]any
+	Hashes      []string   // hash algorithm names in order
+	Schemas     []string   // schema URLs in order
 }
-
-// PayloadRecord holds the content identity of a payload.
-type PayloadRecord struct {
-	Size   int64             `json:"size"`
-	Hashes map[string]string `json:"hashes"`
-}
-
-var ledger *Ledger
 
 var defaultHashes = []string{"blake2b_256", "sha256", "sha1", "md5"}
 
-func NewLedger(w io.Writer, environment map[string]any) error {
-	l := &Ledger{
-		writer:  w,
-		enc:     json.NewEncoder(w),
-		entries: make(chan entryRequest, 256),
-		done:    make(chan struct{}),
-		hashes:  defaultHashes,
-		sigBuf:  make([]byte, 0, 512),
+var defaultSchemas = []string{
+	"https://github.com/buildwarden/buildwarden/schemas/http-open.json",
+	"https://github.com/buildwarden/buildwarden/schemas/http-headers.json",
+	"https://github.com/buildwarden/buildwarden/schemas/http-body.json",
+	"https://github.com/buildwarden/buildwarden/schemas/artifact.json",
+	"https://github.com/buildwarden/buildwarden/schemas/redacted.json",
+}
+
+// NewLedger creates a new binary ledger, writes the header, and starts the
+// serialization loop. The returned Ledger is ready for Open/Checkpoint/Close/Artifact calls.
+func NewLedger(cfg LedgerConfig) (*Ledger, error) {
+	if cfg.Hashes == nil {
+		cfg.Hashes = defaultHashes
+	}
+	if cfg.Schemas == nil {
+		cfg.Schemas = defaultSchemas
 	}
 
 	_, priv, err := ed25519.GenerateKey(rand.Reader)
 	if err != nil {
-		return fmt.Errorf("error generating private key for ledger: %w", err)
+		return nil, fmt.Errorf("generate key: %w", err)
 	}
-	l.key = priv
 
-	if err := l.writeHeader(environment); err != nil {
-		return fmt.Errorf("error writing ledger header: %w", err)
+	hashBlockSize := 0
+	for _, name := range cfg.Hashes {
+		hashBlockSize += hashOutputSize(name)
+	}
+
+	l := &Ledger{
+		key:           priv,
+		writer:        cfg.Writer,
+		entries:       make(chan entryRequest, 256),
+		done:          make(chan struct{}),
+		hashes:        cfg.Hashes,
+		sigSize:       ed25519.SignatureSize, // 64
+		hashBlockSize: hashBlockSize,
+		sigBuf:        make([]byte, 0, 512),
+	}
+
+	if err := l.writeHeader(cfg); err != nil {
+		return nil, err
 	}
 
 	go l.loop()
-	ledger = l
-	return nil
+	return l, nil
 }
 
-func FinishLedger() {
-	if ledger != nil {
-		close(ledger.entries)
-		<-ledger.done
+// PublicKey returns the raw Ed25519 public key bytes.
+func (l *Ledger) PublicKey() ed25519.PublicKey {
+	pub, ok := l.key.Public().(ed25519.PublicKey)
+	if !ok {
+		panic("unexpected key type")
 	}
+	return pub
 }
 
-// PublicCertPEM returns the public key in PEM format.
-func PublicCertPEM() []byte {
-	if ledger == nil {
-		return nil
-	}
-	pub, _ := ledger.key.Public().(ed25519.PublicKey)
-	return pem.EncodeToMemory(&pem.Block{
-		Type:  "ED25519 PUBLIC KEY",
-		Bytes: []byte(pub),
-	})
-}
-
-// Open records an open entry synchronously and returns the entry's signature
-// (which serves as the channel identifier for subsequent checkpoint/close entries).
-func (l *Ledger) Open(metadata map[string]any) string {
-	result := make(chan string, 1)
+// Open writes an open record synchronously and returns the record's raw signature
+// (the channel identifier).
+func (l *Ledger) Open(schemaIndex byte, metadata []byte) []byte {
+	result := make(chan []byte, 1)
 	l.entries <- entryRequest{
 		entry: entryInput{
-			Type:     "open",
-			Metadata: metadata,
+			Type:        RecordOpen,
+			PayloadSize: 0,
+			SchemaIndex: schemaIndex,
+			Metadata:    metadata,
 		},
 		result: result,
 	}
 	return <-result
 }
 
-// Checkpoint records a checkpoint entry (fire-and-forget).
+// Checkpoint writes a checkpoint record (fire-and-forget).
 func (l *Ledger) Checkpoint(
-	openSig string, direction string, payload []byte, metadata map[string]any,
-) {
-	pr := l.computePayload(payload)
-	l.entries <- entryRequest{
-		entry: entryInput{
-			Type:          "checkpoint",
-			OpenSignature: openSig,
-			Direction:     direction,
-			Size:          pr.Size,
-			PreHashed:     pr.Hashes,
-			Metadata:      metadata,
-		},
-	}
-}
-
-// CheckpointHashed records a checkpoint entry with pre-computed hashes (fire-and-forget).
-func (l *Ledger) CheckpointHashed(
-	openSig string, direction string, size int64, hashes map[string]string,
+	openSig []byte, payloadSize int64, hashBlock []byte,
+	schemaIndex byte, metadata []byte,
 ) {
 	l.entries <- entryRequest{
 		entry: entryInput{
-			Type:          "checkpoint",
-			OpenSignature: openSig,
-			Direction:     direction,
-			Size:          size,
-			PreHashed:     hashes,
+			Type:        RecordCheckpoint,
+			OpenSig:     openSig,
+			PayloadSize: payloadSize,
+			HashBlock:   hashBlock,
+			SchemaIndex: schemaIndex,
+			Metadata:    metadata,
 		},
 	}
 }
 
-// Close records a close entry (fire-and-forget).
-func (l *Ledger) Close(openSig string, direction string, payload []byte, metadata map[string]any) {
-	pr := l.computePayload(payload)
-	l.entries <- entryRequest{
-		entry: entryInput{
-			Type:          "close",
-			OpenSignature: openSig,
-			Direction:     direction,
-			Size:          pr.Size,
-			PreHashed:     pr.Hashes,
-			Metadata:      metadata,
-		},
-	}
-}
-
-// CloseHashed records a close entry with pre-computed hashes (fire-and-forget).
-func (l *Ledger) CloseHashed(
-	openSig string, direction string, size int64,
-	hashes map[string]string, metadata map[string]any,
+// Close writes a close record (fire-and-forget).
+func (l *Ledger) Close(
+	openSig []byte, payloadSize int64, hashBlock []byte,
+	schemaIndex byte, metadata []byte,
 ) {
 	l.entries <- entryRequest{
 		entry: entryInput{
-			Type:          "close",
-			OpenSignature: openSig,
-			Direction:     direction,
-			Size:          size,
-			PreHashed:     hashes,
-			Metadata:      metadata,
+			Type:        RecordClose,
+			OpenSig:     openSig,
+			PayloadSize: payloadSize,
+			HashBlock:   hashBlock,
+			SchemaIndex: schemaIndex,
+			Metadata:    metadata,
 		},
 	}
 }
 
-func (l *Ledger) writeHeader(environment map[string]any) error {
-	pub, _ := l.key.Public().(ed25519.PublicKey)
-	certBytes := pem.EncodeToMemory(&pem.Block{
-		Type:  "ED25519 PUBLIC KEY",
-		Bytes: []byte(pub),
-	})
+// Artifact writes an artifact record (fire-and-forget). Closes the channel.
+func (l *Ledger) Artifact(
+	openSig []byte, payloadSize int64, hashBlock []byte,
+	schemaIndex byte, metadata []byte,
+) {
+	l.entries <- entryRequest{
+		entry: entryInput{
+			Type:        RecordArtifact,
+			OpenSig:     openSig,
+			PayloadSize: payloadSize,
+			HashBlock:   hashBlock,
+			SchemaIndex: schemaIndex,
+			Metadata:    metadata,
+		},
+	}
+}
 
-	payload := l.computePayload(certBytes)
+// Finish drains the entry channel and waits for the loop to exit.
+func (l *Ledger) Finish() {
+	close(l.entries)
+	<-l.done
+}
 
-	// Signature input for header: entry_type + size + hashes (no prev_sig)
-	var sigInput []byte
-	sigInput = append(sigInput, []byte("header")...)
-	sigInput = append(sigInput, sizeBytes(payload.Size)...)
-	sigInput = append(sigInput, l.rawHashBytes(payload.Hashes)...)
+// ComputeHashBlock computes the concatenated hash block for the given data.
+func (l *Ledger) ComputeHashBlock(data []byte) []byte {
+	block := make([]byte, 0, l.hashBlockSize)
+	for _, name := range l.hashes {
+		h := newHash(name)
+		h.Write(data)
+		block = h.Sum(block)
+	}
+	return block
+}
 
-	sig, sigRaw, err := l.sign(sigInput)
+// --- internal ---
+
+func (l *Ledger) writeHeader(cfg LedgerConfig) error {
+	pub := l.PublicKey()
+
+	// Build binary prefix
+	var prefix []byte
+	// Magic + version
+	prefix = append(prefix, 'B', 'L', 'D', 'L', 0x01)
+	// Signature scheme (null-terminated)
+	prefix = append(prefix, []byte("ed25519-sha512")...)
+	prefix = append(prefix, 0x00)
+	// Signature size (uint16 big-endian)
+	prefix = binary.BigEndian.AppendUint16(prefix, uint16(l.sigSize))
+	// Hash block size (uint16 big-endian)
+	prefix = binary.BigEndian.AppendUint16(prefix, uint16(l.hashBlockSize))
+	// Public key length (uint16 big-endian)
+	prefix = binary.BigEndian.AppendUint16(prefix, uint16(len(pub)))
+	// Public key bytes
+	prefix = append(prefix, pub...)
+
+	// Sign the binary prefix
+	sig := l.sign(prefix)
+	l.prevSigRaw = sig
+
+	// Encode header CBOR metadata
+	meta := HeaderMeta{
+		Hashes:      cfg.Hashes,
+		Schemas:     cfg.Schemas,
+		Environment: cfg.Environment,
+	}
+	metaBytes, err := cbor.Marshal(meta)
 	if err != nil {
-		return err
+		return fmt.Errorf("encode header metadata: %w", err)
 	}
 
-	l.seq.Store(0)
-	l.prevSigRaw = sigRaw
+	// Write: prefix + signature + metadata-length + metadata
+	var buf []byte
+	buf = append(buf, prefix...)
+	buf = append(buf, sig...)
+	buf = binary.BigEndian.AppendUint32(buf, uint32(len(metaBytes)))
+	buf = append(buf, metaBytes...)
 
-	header := HeaderEntry{
-		EntryType:       "header",
-		Version:         "2.0",
-		Format:          "json",
-		SignatureScheme: "ed25519-sha512",
-		Hashes:          l.hashes,
-		Environment:     environment,
-		Payload:         payload,
-		Signature:       sig,
+	if _, err := l.writer.Write(buf); err != nil {
+		return fmt.Errorf("write header: %w", err)
 	}
-
-	j, err := json.Marshal(header)
-	if err != nil {
-		return fmt.Errorf("error marshalling header: %w", err)
-	}
-	if _, err := l.writer.Write(append(j, '\n')); err != nil {
-		return fmt.Errorf("error writing header: %w", err)
-	}
-
-	l.headerSig = sig
 	return nil
 }
 
@@ -260,126 +269,73 @@ func (l *Ledger) loop() {
 
 	for req := range l.entries {
 		e := req.entry
-		seq := l.seq.Add(1)
+		l.seq.Add(1)
 
-		// Reuse sigBuf: reset length, keep capacity
+		// Build signature input
 		l.sigBuf = l.sigBuf[:0]
+		l.sigBuf = append(l.sigBuf, e.Type)
 		l.sigBuf = append(l.sigBuf, l.prevSigRaw...)
 
-		switch e.Type {
-		case "open":
-			l.sigBuf = append(l.sigBuf, "open"...)
-
-		case "checkpoint", "close":
-			openSigBytes, _ := base64.StdEncoding.DecodeString(e.OpenSignature)
-			l.sigBuf = append(l.sigBuf, openSigBytes...)
-			l.sigBuf = append(l.sigBuf, e.Type...)
-			l.sigBuf = append(l.sigBuf, e.Direction...)
-
-			payload := &PayloadRecord{Size: e.Size, Hashes: e.PreHashed}
-			l.sigBuf = appendSizeBytes(l.sigBuf, payload.Size)
-			l.sigBuf = l.appendRawHashBytes(l.sigBuf, payload.Hashes)
-
-			sig, sigRaw, err := l.sign(l.sigBuf)
-			if err != nil {
-				log.Printf("error signing %s entry: %v", e.Type, err)
-				if req.result != nil {
-					req.result <- ""
-				}
-				continue
-			}
-
-			l.writeEntry(LedgerEntry{
-				EntryType:     e.Type,
-				OpenSignature: e.OpenSignature,
-				Direction:     e.Direction,
-				Payload:       payload,
-				Signature:     sig,
-				Seq:           seq,
-				Timestamp:     time.Now().UTC().Format(time.RFC3339Nano),
-				Metadata:      e.Metadata,
-			})
-			l.prevSigRaw = sigRaw
-			if req.result != nil {
-				req.result <- sig
-			}
-			continue
+		if e.Type != RecordOpen {
+			l.sigBuf = append(l.sigBuf, e.OpenSig...)
 		}
 
-		// Open entry path (no payload)
-		sig, sigRaw, err := l.sign(l.sigBuf)
-		if err != nil {
-			log.Printf("error signing open entry: %v", err)
-			if req.result != nil {
-				req.result <- ""
-			}
-			continue
+		l.sigBuf = binary.BigEndian.AppendUint64(l.sigBuf, uint64(e.PayloadSize))
+
+		if e.PayloadSize != 0 {
+			l.sigBuf = append(l.sigBuf, e.HashBlock...)
 		}
 
-		l.writeEntry(LedgerEntry{
-			EntryType: "open",
-			Signature: sig,
-			Seq:       seq,
-			Timestamp: time.Now().UTC().Format(time.RFC3339Nano),
-			Metadata:  e.Metadata,
-		})
-		l.prevSigRaw = sigRaw
+		sig := l.sign(l.sigBuf)
+
+		// Write record bytes
+		var rec []byte
+		rec = append(rec, e.Type)
+		rec = append(rec, l.prevSigRaw...)
+		if e.Type != RecordOpen {
+			rec = append(rec, e.OpenSig...)
+		}
+		rec = binary.BigEndian.AppendUint64(rec, uint64(e.PayloadSize))
+		if e.PayloadSize != 0 {
+			rec = append(rec, e.HashBlock...)
+		}
+		rec = append(rec, sig...)
+		rec = append(rec, e.SchemaIndex)
+
+		if e.SchemaIndex != SchemaNoMetadata {
+			rec = binary.BigEndian.AppendUint32(rec, uint32(len(e.Metadata)))
+			rec = append(rec, e.Metadata...)
+		}
+
+		if _, err := l.writer.Write(rec); err != nil {
+			log.Printf("error writing ledger record: %v", err)
+		}
+
+		l.prevSigRaw = sig
 		if req.result != nil {
 			req.result <- sig
 		}
 	}
 }
 
-func (l *Ledger) writeEntry(entry LedgerEntry) {
-	if err := l.enc.Encode(entry); err != nil {
-		log.Printf("error writing ledger entry: %v", err)
-	}
-}
-
-func (l *Ledger) sign(input []byte) (string, []byte, error) {
+func (l *Ledger) sign(input []byte) []byte {
 	digest := sha512.Sum512(input)
-	raw := ed25519.Sign(l.key, digest[:])
-	return base64.StdEncoding.EncodeToString(raw), raw, nil
+	return ed25519.Sign(l.key, digest[:])
 }
 
-func (l *Ledger) computePayload(data []byte) *PayloadRecord {
-	hashes := make(map[string]string, len(l.hashes))
-	for _, name := range l.hashes {
-		h := newHash(name)
-		h.Write(data)
-		hashes[name] = hex.EncodeToString(h.Sum(nil))
+func hashOutputSize(name string) int {
+	switch name {
+	case "blake2b_256":
+		return blake2b.Size256
+	case "sha256":
+		return 32
+	case "sha1":
+		return 20
+	case "md5":
+		return 16
+	default:
+		panic("unsupported hash: " + name)
 	}
-	return &PayloadRecord{
-		Size:   int64(len(data)),
-		Hashes: hashes,
-	}
-}
-
-func (l *Ledger) rawHashBytes(hashes map[string]string) []byte {
-	var out []byte
-	for _, name := range l.hashes {
-		b, _ := hex.DecodeString(hashes[name])
-		out = append(out, b...)
-	}
-	return out
-}
-
-func (l *Ledger) appendRawHashBytes(dst []byte, hashes map[string]string) []byte {
-	for _, name := range l.hashes {
-		b, _ := hex.DecodeString(hashes[name])
-		dst = append(dst, b...)
-	}
-	return dst
-}
-
-func sizeBytes(size int64) []byte {
-	buf := make([]byte, 8)
-	binary.LittleEndian.PutUint64(buf, uint64(size))
-	return buf
-}
-
-func appendSizeBytes(dst []byte, size int64) []byte {
-	return binary.LittleEndian.AppendUint64(dst, uint64(size))
 }
 
 func newHash(name string) hash.Hash {
@@ -388,12 +344,55 @@ func newHash(name string) hash.Hash {
 		h, _ := blake2b.New256(nil)
 		return h
 	case "sha256":
-		return sha256.New()
+		return crypto_sha256.New()
 	case "sha1":
-		return sha1.New()
+		return crypto_sha1.New()
 	case "md5":
-		return md5.New()
+		return crypto_md5.New()
 	default:
 		panic("unsupported hash: " + name)
 	}
+}
+
+// HashData computes individual hashes for data, returning them in order.
+// Useful for callers that need both the hash block and individual hex values.
+func HashData(data []byte, hashes []string) [][]byte {
+	result := make([][]byte, len(hashes))
+	for i, name := range hashes {
+		h := newHash(name)
+		h.Write(data)
+		result[i] = h.Sum(nil)
+	}
+	return result
+}
+
+// StreamingHasher computes all configured hashes incrementally.
+type StreamingHasher struct {
+	hashes []hash.Hash
+	size   int64
+}
+
+// NewStreamingHasher creates a hasher for the given algorithm names.
+func NewStreamingHasher(names []string) *StreamingHasher {
+	h := &StreamingHasher{hashes: make([]hash.Hash, len(names))}
+	for i, name := range names {
+		h.hashes[i] = newHash(name)
+	}
+	return h
+}
+
+func (s *StreamingHasher) Write(p []byte) (int, error) {
+	for _, h := range s.hashes {
+		h.Write(p)
+	}
+	s.size += int64(len(p))
+	return len(p), nil
+}
+
+// Finish returns the concatenated hash block and total bytes written.
+func (s *StreamingHasher) Finish() (hashBlock []byte, size int64) {
+	for _, h := range s.hashes {
+		hashBlock = h.Sum(hashBlock)
+	}
+	return hashBlock, s.size
 }
